separate planner from demo script
