+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__/
 
 # C extensions
 *.so
+*.cpp
 
 # Distribution / packaging
 .Python
